feat/refactor/fix: refactored the whole app structure +
