+++ .env
@@ -4,6 +4,9 @@ DB_PASSWORD=tdp1503
 DB_DATABASE=TasksManager
 DB_PORT=3306
 DB_DIALECT=mysql
+
 API_PORT=3001
+
 MIN_PASS_LENGTH=6
-JWT_SECRET=098haeknmrlamsdf0872342l53
+SECRET=098haeknmrlamsdf0872342l53
+EMPTY_FIELD_ERROR=Some required fields are missing
