+++ package.json
@@ -29,12 +29,12 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
     "eject": "react-scripts eject",
-    "pre:start": "npx sequelize-cli db:drop && npx sequelize-cli db:create && npx sequelize-cli db:migrate --migrations-path ./src/api/migrations && npx sequelize-cli db:seed:all --seeders-path ./src/api/seeders",
-    "debug": "npm run pre:start && nodemon ./src/api/index.js"
+    "pre:start": "npx sequelize-cli db:drop && npx sequelize-cli db:create && npx sequelize-cli db:migrate --migrations-path ./server/database/migrations && npx sequelize-cli db:seed:all --seeders-path ./server/database/seeders",
+    "debug": "npm run pre:start && nodemon ./server/index.js"
   },
   "eslintConfig": {
     "extends": [
@@ -56,6 +56,7 @@
   },
   "devDependencies": {
     "eslint": "^8.18.0",
-    "nodemon": "^2.0.16"
+    "nodemon": "^2.0.16",
+    "react-app-rewired": "^2.2.1"
   }
 }
