+++ server/routes/User/index.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const { logIn, signUp } = require('../../controllers/user.controllers');
+const { logIn, signUp } = require('../../database/controllers/user.controllers');
 const { validateLogin, validateSignUp } = require('../../middlewares');
 
 const router = express.Router();
