+++ server/utils/generateJwt.js
@@ -1,13 +1,13 @@
 const jwt = require('jsonwebtoken');
 
-const secret = process.env.JWT_SECRET;
+const { SECRET } = process.env;
 
 const jwtConfig = {
   algorithm: 'HS256',
 };
 
 module.exports = (payload) => {
-  const token = jwt.sign({ data: payload }, secret, jwtConfig);
+  const token = jwt.sign({ data: payload }, SECRET, jwtConfig);
 
   return token;
 };
